Move regex helpers to new module
